+++ src/main/java/com/code/challenge/Application.java
@@ -1,0 +1,23 @@
+package com.code.challenge;
+
+import jakarta.annotation.PostConstruct;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+import java.util.Date;
+import java.util.TimeZone;
+
+@SpringBootApplication
+@EnableScheduling
+public class Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+
+	@PostConstruct
+	public void fixJvmTimeZone(){
+		TimeZone.setDefault(TimeZone.getTimeZone("UTC"));
+	}
+}
