+++ src/main/java/com/code/challenge/api/EventSummaryApiModel.java
@@ -1,0 +1,6 @@
+package com.code.challenge.api;
+
+import java.math.BigDecimal;
+import java.sql.Timestamp;
+
+public record EventSummaryApiModel(String eventType, BigDecimal eventValue, Timestamp lastUpdatedAt) { }
