+++ src/main/java/com/code/challenge/api/MachineApiModel.java
@@ -1,0 +1,3 @@
+package com.code.challenge.api;
+
+public record MachineApiModel(String machineId) { }
