+++ src/main/java/com/code/challenge/api/MachineController.java
@@ -1,0 +1,28 @@
+package com.code.challenge.api;
+
+import com.code.challenge.models.Machine;
+import com.code.challenge.models.MachineRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@RestController
+@RequestMapping(path="/machines")
+public class MachineController {
+    @Autowired
+    private MachineRepository machineRepository;
+
+    @GetMapping(path="/list")
+    public List<MachineApiModel> getAll() {
+        Iterable<Machine> machinesList = machineRepository.findAll();
+        List<MachineApiModel> machinesApi = new ArrayList();
+        machinesList.forEach(machine -> {
+            machinesApi.add(new MachineApiModel(machine.getMachineId()));
+        });
+        return machinesApi;
+    }
+}
