+++ src/main/java/com/code/challenge/api/SessionApiModel.java
@@ -1,0 +1,5 @@
+package com.code.challenge.api;
+
+import java.sql.Timestamp;
+
+public record SessionApiModel(String sessionId, String machineId, Timestamp startedAt, Timestamp finishedAt) { }
