+++ src/main/java/com/code/challenge/api/SessionSummaryApiModel.java
@@ -1,0 +1,5 @@
+package com.code.challenge.api;
+
+import java.util.List;
+
+public record SessionSummaryApiModel(String sessionId, String machineId, List<EventSummaryApiModel> eventSummary) { }
