--- src/main/java/com/example/template/api/EventSummaryApiModel.java
@@ -1,6 +1,0 @@
-package com.example.template.api;
-
-import java.math.BigDecimal;
-import java.sql.Timestamp;
-
-public record EventSummaryApiModel(String eventType, BigDecimal eventValue, Timestamp lastUpdatedAt) { }
