--- src/main/java/com/example/template/api/MachineApiModel.java
@@ -1,3 +1,0 @@
-package com.example.template.api;
-
-public record MachineApiModel(String machineId) { }
