--- src/main/java/com/example/template/api/SessionApiModel.java
@@ -1,5 +1,0 @@
-package com.example.template.api;
-
-import java.sql.Timestamp;
-
-public record SessionApiModel(String sessionId, String machineId, Timestamp startedAt, Timestamp finishedAt) { }
