--- src/main/java/com/example/template/api/SessionSummaryApiModel.java
@@ -1,5 +1,0 @@
-package com.example.template.api;
-
-import java.util.List;
-
-public record SessionSummaryApiModel(String sessionId, String machineId, List<EventSummaryApiModel> eventSummary) { }
