+++ src/main/resources/application.properties
@@ -5,3 +5,5 @@ spring.datasource.password=codechallengepassword
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 #spring.jpa.show-sql: true
 sqs.endpoint.configuration=http://localhost:4566
+# Important to not kill ongoing requests
+server.shutdown=graceful
